Naive lexer and parser
This implements a fairly naive lexer and parser, and has an inefficient
pipeline for converting to Python.  All the operations are first parsed
into an internal byte code representation in Rust, and then this entire
vector is converted to Python objects (a slightly different byte-code
representation).

Gate declarations are not currently supported.  Mathematical-expression
parsing is also not yet implemented.

Both the lexer and parser are hand-written.  This was because this first
iteration of the project was for my own practice and experience; I
deliberately didn't use existing frameworks to learn a bit more about
how lexers and parsers can be implemented, and what the problems are
with the naive ways I might originally write them.

The lexer is not far off a simple hardcoded DFA.  Parsing is simple
recursive descent.