Convert to BufRead-based TokenStream
This is an important step for two developments of the library:
abstraction allowing for loading from files and strings interchangeably,
and exposing the bytecode stream as a custom Python iterator, rather
than needing to eagerly construct the entire byte code first.

This involved switching to iteration over bytes rather than characters,
but since OpenQASM 2 only allows ASCII anyway, this shouldn't be a big
deal.  The extra overhead from all the additional error checking in the
naive conversions of `next_byte` and `peek_byte` is not huge; the cost
of parsing and the construction of the Python object at the end is
larger.  This could be improved by being smarter about where errors are
actually checked, though.