Convert output to lazy Python iterators
This naive approach seems to have come with a slight slow down, but
should reduce the intermediate memory usage of the library and allow
Python `KeyboardInterrupt` to process (effectively) during the lexing
and parsing stages as well.  Precisely, the interrupt will only trigger
when returning to Python space, but we no longer need to parse the whole
program in Rust space before rentering Python now.