Fix packaging and tox
Using `attr: qiskit_qasm2.__version__` had problems for installing the
package from an sdist; the package needed to be available before it was
actually built in this situation.  Instead, we use a common file for a
single source of truth, and load this both at import and in the setup
file.

Some of the tox configuration had minor copy-paste errors in it, and the
the shared environments were no longer achieving the desired effect in
tox 4 due to better isolation between environment configurations
changing.