Add internal documentation
This adds all the missing internal documentation of all the Rust parts
of the code, and some of the Python components that are not part of the
public interface.  I was a bit lazy when I was originally writing the
code, because I was just trying a lot of different things out as I
learned more about what I was doing.