Fix iteration with no-op statments
Fixing broadcasts of zero size caused an issue with the iteration logic,
since now not every apparent statement in the OpenQASM file produced one
or more bytecode statements for the Python interpreter.  Previously,
this was the condition for the iterator to terminate, but this caused
no-op statements to end the parsing prematurely.  Instead, we now
explicitly return how many statements were emitted from each statement
step, and the main parser `next` method simply loops until it either
reaches the end of the file (in which case it now explicitly returns
`None`) or returns some number of emitted instructions.