Add tests of correct OpenQASM 2 programs
This so far does not include all the tests of the behaviour in the face
of _incorrect_ OpenQASM programs, but that will come later.  The two
main strands of tests are integration tests via each of the examples in
the OpenQASM 2 arXiv paper, and unit testing of individual statement
types.

Several bugfix commits preceding this one were uncovered during the
writing of these tests.