Add tests of expression parsing
These tests turned up significant issues in the current method of using
Terra's `ParameterExpression` to represent the lazily-applied gate
bodies, which are generally unsolvable within the current method; Terra
is simply missing methods to apply `sqrt` to `ParameterExpression`s, or
to raise them to powers of each other.

Most likely this can be solved by a rewrite of the Python output of the
expression components into a Python tree structure, which the
`_DefinedGate` structure will evaluate on definition.