Convert expression handling to subordinate bytecode
The previous form attempted to use Qiskit's `Parameter` objects to build
the gate definitions lazily, but they could not support square roots or
powers.  They also necessitated storing an extra `QuantumCircuit`
instance in order to do the binding.  Since we already have most of the
bytecode and ability to interpret it in place, it was easier to have
complete expression handling just by swapping over the expression
representation to pass up completely to Python, and then store a mini
bytecode interpreter in the Python gate definitions.