Avoid unnecessary allocations during lexing
All valid OpenQASM 2 tokens can only occur on a single line.  Since we
read through the stream line-by-line, we always have the relevant bytes
for each token stored in our buffer already, and we only copy them into
our context object if we don't already have them stored.  This means we
only actually need to copy the bytes over if we _don't_ already have
them stored, and we can check that just by hashing the relevant part of
the byte stream, rather than copying and heap-allocating a string for
every token we see.

We still create a string when storing it in our context object, but this
happens infrequently for each token.