Add support for overriding instructions
This supports users overriding the objects that get emitted for any
given quantum instruction in an OpenQASM 2 input.  This includes both
"custom" instructions that still need to be defined by the OpenQASM 2
file (with a matching signature, but gate bodies are ignored), and
"builtins", which are automatically available, but will not raise an
error if a compatible signature declaration is given.

This allows us to expose a `QISKIT_CUSTOM_INSTRUCTIONS` tuple, which
contains all the `qelib1.inc` additions that Terra currently makes as
built-ins.  This is not _quite_ the same behaviour as Terra (since it
technically still needs to see a declaration for each custom gate to
make it available), but in practice will be largely the same.

This implementation retains type safety in the custom instructions,
unlike Terra.  Terra will ignore whatever the signature of the
definition is, and still attempt to use the same custom objects.  This
can cause them to raise exceptions due to being constructed incorrectly.