Add `strict` mode and trailing commas/semicolons in non-strict
It's sometimes convenient for programmatic generation / consistent
styling of OQ2 to allow trailing commas in parameter and qubit lists,
and typing an extra semicolon after gate bodies is something I commonly
do by mistake.  This relaxes the parser to allow those cases.

In addition, this adds a `strict` mode, to disable the same relaxations,
and to require that the input contain and start with a suitable
version statement.  This mode is in line with how the Qiskit parser
handles things; it is tight to the letter of the specification.