Add support for custom classical functions
This adds the ability to pass custom Python callable functions into the
parser to supplement the "scientific calculator" functionality described
by the paper.  This makes it possible to implement Qiskit's
classical-component extensions to the langauge, while keeping the main
parser correct to the letter of the specification.

The allowed callables are the set of functions that take in some fixed
number of floating-point values, and return a single floating-point
value.  OpenQASM 2 doesn't have the type system to represent anything
more than that, and if I added one, the language wouldn't really be
OpenQASM 2 any more.

The tuple `QISKIT_CUSTOM_CLASSICAL` is added for use as a Qiskit
compatibility mode.  With this in place, it is possible to use this
module to completely replace `QuantumCircuit.from_qasm_str` and
`from_qasm_file` when called with certain inputs, and the full interface
is far more extensible and configurable, and significantly faster.