Improve error messages from lexer
This is a complete refactor of how errors are handled in the lexer,
overloading the `TokenContext` interner to store error messages and pass
them through to the parser, which is actually capable of emitting
them.  As a pleasant side effect, this reduces the amount of data passed
around during the lexing stage.  This significantly improves the quality
of error messages returned by the lexer.

In retrospect, it may be worth investigating the performance of keeping
the special `TokenType::Error` type, or whether it might be more
appropriate to have the `TokenStream` iterate over a `Result<Token, _>`
type, forcing that the error case is handled everywhere.