Refactor error messages to use generic position
For the most part, the error messages should be tied to some sort of
"position" or "span" rather than specifically a token.  This allows the
lexer to use the same error-message builders in cases where it can't
build a complete token as well.

For now, I just use a single starting `Position` because I don't really
have the time to emit complete diagnostics including underlining the
whole problematic span (or even more detailed things like underlining
the original declaration in type errors), and all I really _need_ to
output is a pointer to the start of the offence.